# api/app/controllers/schemas.py
from datetime import datetime

from pydantic import BaseModel, ConfigDict, EmailStr, Field, field_validator

from ..persistence.models import PhysicalBookStatus, UserRole


def _validate_isbn13(isbn: str) -> str:
    """`Book.isbn` is a String(13) primary key, so only well-formed ISBN-13 gets in."""
    if len(isbn) != 13 or not isbn.isdigit():
        raise ValueError("isbn must be 13 digits")
    # Check digit: digits weighted 1,3,1,3,... must add up to a multiple of 10.
    total = sum(int(digit) * (1 if index % 2 == 0 else 3) for index, digit in enumerate(isbn))
    if total % 10 != 0:
        raise ValueError("isbn has an invalid check digit")
    return isbn


class LibraryBase(BaseModel):
    name: str
    address: str
    state: str
    city: str
    hours: str | None = None
    phone: str | None = None
    email: str | None = None
    website: str | None = None


class LibraryCreate(LibraryBase):
    pass


class LibraryUpdate(BaseModel):
    name: str | None = None
    address: str | None = None
    state: str | None = None
    city: str | None = None
    hours: str | None = None
    phone: str | None = None
    email: str | None = None
    website: str | None = None


class LibraryOut(LibraryBase):
    model_config = ConfigDict(from_attributes=True)
    id: int


class AuthorOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    id: int
    name: str


class GenreOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    id: int
    name: str


class BookBase(BaseModel):
    isbn: str
    title: str
    language: str
    pages: int | None = None
    synopsis: str | None = None


class BookOut(BookBase):
    model_config = ConfigDict(from_attributes=True)
    authors: list[AuthorOut] = []
    genres: list[GenreOut] = []


class BookCreate(BookBase):
    author_ids: list[int] = []
    genre_ids: list[int] = []

    @field_validator("isbn")
    @classmethod
    def check_isbn(cls, value: str) -> str:
        return _validate_isbn13(value)


class BookUpdate(BaseModel):
    title: str | None = None
    language: str | None = None
    pages: int | None = None
    synopsis: str | None = None
    # Sending a list replaces the whole association; omitting it leaves it as is.
    author_ids: list[int] | None = None
    genre_ids: list[int] | None = None


class AuthorCreate(BaseModel):
    name: str = Field(min_length=1)


class AuthorUpdate(BaseModel):
    name: str | None = Field(default=None, min_length=1)


class GenreCreate(BaseModel):
    name: str = Field(min_length=1)


class GenreUpdate(BaseModel):
    name: str | None = Field(default=None, min_length=1)


class PhysicalBookOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    id: int
    isbn: str
    library_id: int
    status: PhysicalBookStatus


class PhysicalBookCreate(BaseModel):
    isbn: str
    library_id: int

    @field_validator("isbn")
    @classmethod
    def check_isbn(cls, value: str) -> str:
        return _validate_isbn13(value)


class PhysicalBookStatusUpdate(BaseModel):
    status: PhysicalBookStatus


class LibraryAvailability(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    library: LibraryOut
    available_copies: int
    # A representative physical book id from this library, ready to be reserved.
    physical_book_id: int


class BookAvailability(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    book: BookOut
    libraries: list[LibraryAvailability]


class UserOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    id: int
    email: str
    name: str
    language: str
    role: UserRole
    library_id: int | None = None


class UserCreate(BaseModel):
    email: EmailStr
    password: str = Field(min_length=8)
    name: str
    language: str = "es"


class UserStaffCreate(BaseModel):
    """Alta de personal (`librarian`/`sysadmin`); solo para sysadmins."""

    email: EmailStr
    password: str = Field(min_length=8)
    name: str
    language: str = "es"
    role: UserRole
    library_id: int | None = None


class UserUpdate(BaseModel):
    name: str | None = None
    language: str | None = None
    password: str | None = Field(default=None, min_length=8)
    # Only a sysadmin may send these two; anyone else gets a 403.
    role: UserRole | None = None
    library_id: int | None = None


class LoginRequest(BaseModel):
    email: EmailStr
    password: str


class TokenOut(BaseModel):
    access_token: str
    token_type: str = "bearer"
    expires_in: int


class ReservationCreate(BaseModel):
    physical_book_id: int
    expires_at: datetime


class ReservationOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    id: int
    user_id: int
    physical_book_id: int
    reserved_at: datetime
    expires_at: datetime
    picked_up: bool