# api/openapi.yml
openapi: 3.0.3
info:
  title: BookUp API
  description: >
    API del MVP de BookUp: catálogo de libros, sedes de biblioteca, reservas de
    ejemplares físicos y usuarios. Generado a partir del código en
    `api/app/controllers/` (routers FastAPI) y `api/app/controllers/schemas.py`.
  version: "0.1.0"
servers:
  - url: http://localhost:8000
    description: Desarrollo local
...
paths:
...
  /books:
    get:
      tags: [catalog]
      summary: Listar todos los libros del catálogo
      operationId: listBooks
      responses:
        "200":
          description: Lista de libros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookOut"
    post:
      tags: [catalog]
      summary: Crear un libro en el catálogo
      description: >
        Requiere `librarian` o `sysadmin`. El `isbn` tiene que ser un ISBN-13
        válido (13 dígitos y dígito verificador correcto), si no da 422.
        `author_ids`/`genre_ids` referencian autores y géneros ya existentes:
        si alguno no existe, 404.
      operationId: createBook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookCreate"
      responses:
        "201":
          description: Libro creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /books/search:
    get:
      tags: [catalog]
      summary: Buscar libros por texto libre
      description: >
        Búsqueda por título/ISBN/autor usando `ILIKE` (placeholder; en la
        arquitectura target se reemplaza por OpenSearch).
      operationId: searchBooks
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Libros que matchean la búsqueda
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookOut"
        "422":
          $ref: "#/components/responses/ValidationError"
  /books/{isbn}:
    get:
      tags: [catalog]
      summary: Obtener un libro por ISBN
      operationId: getBook
      parameters:
        - $ref: "#/components/parameters/IsbnPath"
      responses:
        "200":
          description: Libro encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOut"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [catalog]
      summary: Actualizar un libro
      description: >
        Requiere `librarian` o `sysadmin`. Actualización parcial; solo se
        aplican los campos enviados. Enviar `author_ids`/`genre_ids`
        **reemplaza** la lista completa (mandar `[]` la vacía); omitirlos la
        deja como estaba.
      operationId: updateBook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IsbnPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Libro actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [catalog]
      summary: Eliminar un libro del catálogo
      description: >
        Requiere `librarian` o `sysadmin`. Falla con 409 si el libro todavía
        tiene ejemplares físicos en alguna sede.
      operationId: deleteBook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IsbnPath"
      responses:
        "204":
          description: Libro eliminado
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /books/{isbn}/availability:
    get:
      tags: [catalog]
      summary: Disponibilidad de un libro por sede
      description: >
        Agrupa los ejemplares físicos (`PhysicalBook`) disponibles del libro
        por sede, e informa un `physical_book_id` representativo listo para
        reservar en cada sede.
      operationId: getBookAvailability
      parameters:
        - $ref: "#/components/parameters/IsbnPath"
      responses:
        "200":
          description: Disponibilidad por sede
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookAvailability"
        "404":
          $ref: "#/components/responses/NotFound"
  /authors:
    get:
      tags: [authors]
      summary: Listar autores
      operationId: listAuthors
      responses:
        "200":
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthorOut"
    post:
      tags: [authors]
      summary: Crear un autor
      description: >
        Requiere `librarian` o `sysadmin`. `Author.name` no es único en el
        modelo: se permiten homónimos (dos autores distintos pueden llamarse
        igual) y la identidad es el `id`, así que crear el mismo nombre dos
        veces devuelve dos autores distintos, no un 409.
      operationId: createAuthor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorCreate"
      responses:
        "201":
          description: Autor creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /authors/{author_id}:
    get:
      tags: [authors]
      summary: Obtener un autor por id
      operationId: getAuthor
      parameters:
        - $ref: "#/components/parameters/AuthorIdPath"
      responses:
        "200":
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorOut"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [authors]
      summary: Actualizar un autor
      description: Requiere `librarian` o `sysadmin`.
      operationId: updateAuthor
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorUpdate"
      responses:
        "200":
          description: Autor actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [authors]
      summary: Eliminar un autor
      description: >
        Requiere `librarian` o `sysadmin`. Falla con 409 si el autor todavía
        está asociado a algún libro.
      operationId: deleteAuthor
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorIdPath"
      responses:
        "204":
          description: Autor eliminado
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /genres:
    get:
      tags: [genres]
      summary: Listar géneros
      operationId: listGenres
      responses:
        "200":
          description: Lista de géneros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GenreOut"
    post:
      tags: [genres]
      summary: Crear un género
      description: >
        Requiere `librarian` o `sysadmin`. A diferencia de los autores,
        `Genre.name` **sí** es único: un nombre repetido da 409.
      operationId: createGenre
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenreCreate"
      responses:
        "201":
          description: Género creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /genres/{genre_id}:
    get:
      tags: [genres]
      summary: Obtener un género por id
      operationId: getGenre
      parameters:
        - $ref: "#/components/parameters/GenreIdPath"
      responses:
        "200":
          description: Género encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreOut"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [genres]
      summary: Actualizar un género
      description: Requiere `librarian` o `sysadmin`. 409 si el nombre ya lo usa otro género.
      operationId: updateGenre
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/GenreIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenreUpdate"
      responses:
        "200":
          description: Género actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [genres]
      summary: Eliminar un género
      description: >
        Requiere `librarian` o `sysadmin`. Falla con 409 si el género todavía
        está asociado a algún libro.
      operationId: deleteGenre
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/GenreIdPath"
      responses:
        "204":
          description: Género eliminado
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT emitido por `POST /auth/login`, enviado en el header
        `Authorization` con el prefijo `Bearer`.
  parameters:
    IsbnPath:
      name: isbn
      in: path
      required: true
      schema:
        type: string
...
    AuthorIdPath:
      name: author_id
      in: path
      required: true
      schema:
        type: integer
    GenreIdPath:
      name: genre_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    NotFound:
      description: El recurso solicitado no existe
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    Unauthorized:
      description: Falta el token, está vencido o es inválido
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    Forbidden:
      description: El usuario está autenticado pero no tiene permiso para esta operación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    Conflict:
      description: La operación entra en conflicto con el estado actual del recurso
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    ValidationError:
      description: Error de validación del payload/parámetros (FastAPI/Pydantic)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
  schemas:
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
      example:
        detail: "Book 9780000000000 not found"
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string
    LibraryBase:
      type: object
      required: [name, address, state, city]
      properties:
        name:
          type: string
        address:
          type: string
        state:
          type: string
        city:
          type: string
        hours:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
...
    LibraryOut:
      allOf:
        - $ref: "#/components/schemas/LibraryBase"
        - type: object
          required: [id]
          properties:
            id:
              type: integer
    AuthorOut:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    AuthorCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    AuthorUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          nullable: true
    GenreOut:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    GenreCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    GenreUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          nullable: true
    BookBase:
      type: object
      required: [isbn, title, language]
      properties:
        isbn:
          type: string
        title:
          type: string
        language:
          type: string
        pages:
          type: integer
          nullable: true
        synopsis:
          type: string
          nullable: true
    BookOut:
      allOf:
        - $ref: "#/components/schemas/BookBase"
        - type: object
          properties:
            authors:
              type: array
              items:
                $ref: "#/components/schemas/AuthorOut"
              default: []
            genres:
              type: array
              items:
                $ref: "#/components/schemas/GenreOut"
              default: []
    BookCreate:
      allOf:
        - $ref: "#/components/schemas/BookBase"
        - type: object
          properties:
            isbn:
              type: string
              description: ISBN-13 válido (13 dígitos y dígito verificador correcto).
              pattern: "^[0-9]{13}$"
            author_ids:
              type: array
              items:
                type: integer
              default: []
            genre_ids:
              type: array
              items:
                type: integer
              default: []
    BookUpdate:
      type: object
      description: >
        Actualización parcial. `author_ids`/`genre_ids` reemplazan la lista
        completa cuando se envían.
      properties:
        title:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        pages:
          type: integer
          nullable: true
        synopsis:
          type: string
          nullable: true
        author_ids:
          type: array
          items:
            type: integer
          nullable: true
        genre_ids:
          type: array
          items:
            type: integer
          nullable: true
...
    LibraryAvailability:
      type: object
      required: [library, available_copies, physical_book_id]
      properties:
        library:
          $ref: "#/components/schemas/LibraryOut"
        available_copies:
          type: integer
        physical_book_id:
          type: integer
          description: Id representativo de un ejemplar físico disponible en esta sede, listo para reservar.
    BookAvailability:
      type: object
      required: [book, libraries]
      properties:
        book:
          $ref: "#/components/schemas/BookOut"
        libraries:
          type: array
          items:
            $ref: "#/components/schemas/LibraryAvailability"
...